+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   (a possible issue on Windows).
 * `jyutping_to_yale` and `parse_jyutping` now return a null value
   (rather than raise an error) when the input is null.
+* The word segmentation function `segment` now strips all whitespace
+  from the input unsegmented string before segmenting it.
 
 ### Security